Improve NTP handling
Instead of trying to detect time not properly set, use the same
technique as for captive portal detection: fallback to the system DNS
over UDP. Only allow this for 10 minutes after daemon startup so we
can't get a downgrade attack by just blocking 443.

In addition to system DNS, NextDNS anycast IP is added to the list for
the case when system DNS cannot be detected.