Publish Debian/RPM packages to S3
This is a first tentative to publish Debian and RPM packages to S3
instead of Bintray. Currently, only the latest version is available.
In the future, this could be changed by resyncing the S3 bucket in the
other direction. It requires secret variable S3_REPO_KEY and
S3_REPO_SECRET.

The install script has been updated to use the new repositories. Both
Debian and RPM repositories are signed, using the key stored in
GPG_KEY. For RPM, we only sign the repository metadata. This is enough
as it contains checksum of the databases which contain checksums of
the packages.

The repository also contains the public key as well as the .repo for
RPM distributions.

Next step is to allow existing users to auto-upgrade to the new
repositories.