Fix detection of Upstart init (#460)
initctl binary is also present for SysV system. Debian not running
systemd would be misdetected as Upstart. The usual way to detect if
Upstart is installed and running is to query `initctl version`.

This fix has been tested on Debian 8 without systemd, Ubuntu 14 with
Upstart not running and Ubuntu 14 with Upstart running.