Make Debian/RPM packages update the repository location
The postinstall scripts are tweaked to update the repository if the
old one is found. For Debian users, the GPG key is also installed at
the expected location. RPM users will have to accept the new key.

This should make the upgrade of existing users transparent without
tampering with one user's system too much. An alternative would have
been to ship the GPG key and .repo with the packages. This attempt was
made at 40c306576d3b, but it is more invasive.

The current code could be removed once bintray is out of service.