Publish Debian/RPM packages to S3 (#463)
* Publish Debian/RPM packages to S3

This is a first tentative to publish Debian and RPM packages to S3
instead of Bintray. Currently, only the latest version is available.
In the future, this could be changed by resyncing the S3 bucket in the
other direction. It requires secret variable S3_REPO_KEY and
S3_REPO_SECRET.

The install script has been updated to use the new repositories. Both
Debian and RPM repositories are signed, using the key stored in
GPG_KEY. For RPM, we only sign the repository metadata. This is enough
as it contains checksum of the databases which contain checksums of
the packages.

The repository also contains the public key as well as the .repo for
RPM distributions.

Next step is to allow existing users to auto-upgrade to the new
repositories.

* Make Debian/RPM packages update the repository location

The postinstall scripts are tweaked to update the repository if the
old one is found. For Debian users, the GPG key is also installed at
the expected location. RPM users will have to accept the new key.

This should make the upgrade of existing users transparent without
tampering with one user's system too much. An alternative would have
been to ship the GPG key and .repo with the packages. This attempt was
made at 40c306576d3b, but it is more invasive.

The current code could be removed once bintray is out of service.

* Temporarily disable release to Bintray/OpenWRT

It is safer to make a first release without uploading to Bintray. We
do not want to bother OpenWRT people by also making a PR. Once the new
repository is considered valid, we can revert this commit.