For UDM Pro, divert traffic to NextDNS using addrtype instead of RFC1918 (#507)
In commit 8ee250abaa16, to divert DNS traffic to NextDNS, the traffic
was selected using RFC1918 address as a destination. We can do that
more cleanly by specifically asking Netfilter to match any local
address.

Note this matches comment of 8ee250abaa16 "Make UDM only capturing DNS
traffic directed at the router", but this does not have the same
effect. Notably, it works if the router is not using RFC 1918 but it
would also work if traffic is targeting the public IP address of the
external interface, which may not be what we expect. However, the
FORWARD rules should still apply.