proxy: limit the number of concurrent inflight requests (#521)
With a default limit of 256 inflight requests, a latency of 3ms to
NextDNS, asking `dnsperf` to run at 100 QPS, there is no change in
latency (which is expected).

The limit can be adjusted using `-max-inflight-requests`. Ideally,
this could be tuned depending on the memory available on the system,
but it seems complex to do that portably in Go. I think 256 is a fine
limit for most systems. We could have a different default limit for
smaller archs (like mips) if we wanted to do some kind of "detection".