Make snapshot action work when we are not rs
Due to security concerns, actions running on a pull_request workflow
do not have access to secrets. The workflow need to be cut in two: the
untrusted part will test and build and upload the result as an
artifact. The trusted part will take the artifact and upload it.

Downloading the artifact is a bit complex until
actions/download-artifact is able to do that for us (see
https://github.com/actions/download-artifact/issues/60). The code
snippet is stolen from
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/.

It is a bit a pain to test as it needs to be merged in master before
being able to test. I hope to get this right at the first try.
Otherwise, I'll try to test in my own fork.