Add integration tests using Fedora/Ubuntu/Debian cloud images
The tests are currently quite simple:

 - install NextDNS
 - run some `nextdns` command
 - activate NextDNS
 - resolve some random names

The tests rely on cloud images with cloud-init enabled. They also
rely on ability to use 9P to pass the test script (this could be done
differently) and to retrieve the result (a bit harder to do
differently, but we could pass the result through a serial console).
It should be easy to add other cloud-init-enabled Linux but maybe it
could be extended to BSDs (see https://bsd-cloud-image.org/).

One drawback is that if there is a failure outside of the test script,
the VMs will just linger idle for 15 minutes. The VMs are running at
slow speed because nested virtualization is not available inside
GitHub environment. Otherwise, the tests would run in less than a
minute.

Example of result: https://github.com/vincentbernat/nextdns/actions/runs/979162589