Embed Let's Encrypt root certificates
This avoids loading all the system root certificates. On my system,
3MB of memory are saved doing this. However, this comes with a high
risk as if we need to switch away from Let's Encrypt without careful
planning, not-up-to-date clients won't be able to connect at all.

This could be solved by:
 - using a flag to not force use of root system certificates
 - load root system certificates in case of bad certificate error

I think the second option is doable. However, are 3MB of savings worth
the complexity? I thought it would have been possible to save more.