Add some debug for snapshot tests (#549)
For some reason, the status of the tests are not added to the PR. It
was working on my repository. Let's add some debug to understand why
this happens. Notably, owner/repo may not be set correctly.

This needs to be merged as is to be able to test: we cannot test this
workflow in a PR.