Use system certificates in addition to LE certificates
The first idea was to fallback to system certificates if we receive a
"bad certificate" error, however, this brings additional complexity
and some edge cases are difficult to handle, like when the first
error is a transient connection error.

This commit changes the goal of the PR: instead of using less memory,
we use the same amount, but we ensure the LE certificates are present.
Notably, this would fix issues like #500 and #546.

This may be removed in the future if we move away from LE.