install: normalize umask value when installing (#596)
If user's umask is 0077, some files may be installed and readable as
root only. apt is running signature check as an unprivileged user and
therefore needs the GPG key to be world-readable. We may run into
other similar issues in the future, so it seems better to set umask to
0022 instead of trying to fix permissions everywhere.

Fix #595.