github: fix architecture mapping (#617)
There was a change in how Debian packages are built. The
"Architecture" field of the generated Debian packages are now a bit
different and we need to adapt.

Instead of trying to map Go architectures to Debian architectures then
to other architectures, I am mapping Go architectures to
Debian/Alpine/RPM architectures.

On Debian:
 - armv5 → nothing
 - armv6 → armel
 - armv7 → armhf (see https://www.debian.org/releases/stable/armhf/release-notes/ch-whats-new.en.html#idm120)

On Raspbian, this was not the same as their port was using hard-float
ABI on ARMv6, but the armel version should work fine. And Raspbian is
a bit a thing of the past I think.

On Alpine:
 - armv6 → armhf
 - armv7 → armv7

On RPM distributions, ARMv7 distribution is armhfp. I am also fixing
the official name used for x86_64. Hopefully, it seems yum/dnf are
more lenient on the names.

For MIPS, it is now using non-official names, so I think this cannot
be fixed right away. Whatever we choose hard or softfloat, the
architecture name contains the choice:

```
 09:20 ❱ dpkg -I ./nextdns_1.37.4_linux_mips64_hardfloat.deb| grep Architecture
 Architecture: mips64hardfloat
 09:20 ❱ dpkg -I ./nextdns_1.37.4_linux_mips64_softfloat.deb| grep Architecture
 Architecture: mips64softfloat
```

With a bit of luck, this does not matter much and I am only fixing the
generated dists file to remove the information.