release: use an armored key for RPM repositories
For Debian repositories, the key needs to *not* be armored as this is
only supported since APT 1.4 (so Debian Stretch and Ubuntu Trusty
needs a non-armored key).

As reported on #558, on RPM distributions, the key is expected to be
armored and some implementations only work with an armored key.

So, we generate both key formats, we keep `nextdns.gpg` for Debian and
we use `nextdns-armored.gpg` for RPM.

Fix #558.