release: use centos:8 image but modify repo list to use vault.centos.org (#663)
CentOS 8 is out of support since end of 2021 and the repositories have
been moved to vault.centos.org. The Docker images have not been
updated. We patch the repository definitions to use vault.centos.org.

An alternative would be to use Fedora as a base instead, but it seems
safer to use CentOS 8 since it works and we know the repositories
created by the version of createrepo is CentOS work.